+++ packages/keyring-snap-bridge/src/versions.test.ts
@@ -1,35 +1,35 @@
-import { KeyringVersion } from '@metamask/keyring-internal-api';
+import { KeyringInternalFeature } from '@metamask/keyring-internal-api';
 import { isValidSemVerVersion, type SemVerVersion } from '@metamask/utils';
 
 import {
-  getKeyringVersionFromPlatform,
-  PLATFORM_VERSION_TO_KEYRING_VERSION,
+  getKeyringFeaturesFromPlatform,
+  PLATFORM_VERSION_TO_KEYRING_FEATURES,
 } from './versions';
 
-describe('getKeyringVersionFromPlatform', () => {
-  it('gets the keyring version v1 as default', () => {
+describe('getKeyringFeaturesFromPlatform', () => {
+  it('gets an empty feature set as default', () => {
     const isSupportedVersion = (): boolean => false;
 
-    expect(getKeyringVersionFromPlatform(isSupportedVersion)).toBe(
-      KeyringVersion.V1,
+    expect(getKeyringFeaturesFromPlatform(isSupportedVersion)).toStrictEqual(
+      new Set(),
     );
   });
 
-  it('gets the keyring version v2 if the platform version is 7.0.0', () => {
+  it('gets the UseOrigin feature if the platform version is 7.0.0', () => {
     const isSupportedVersion = (version: SemVerVersion): boolean => {
       return version === '7.0.0';
     };
 
-    expect(getKeyringVersionFromPlatform(isSupportedVersion)).toBe(
-      KeyringVersion.V2,
+    expect(getKeyringFeaturesFromPlatform(isSupportedVersion)).toStrictEqual(
+      new Set([KeyringInternalFeature.UseOrigin]),
     );
   });
 });
 
-describe('PLATFORM_VERSION_TO_KEYRING_VERSION', () => {
-  it.each(PLATFORM_VERSION_TO_KEYRING_VERSION)(
+describe('PLATFORM_VERSION_TO_KEYRING_FEATURES', () => {
+  it.each(PLATFORM_VERSION_TO_KEYRING_FEATURES)(
     'is a valid semver version: %s',
-    // `PLATFORM_VERSION_TO_KEYRING_VERSION` items are flattened by `it.each`, no need
+    // `PLATFORM_VERSION_TO_KEYRING_FEATURES` items are flattened by `it.each`, no need
     // to destructure the tuple here.
     (version) => {
       expect(isValidSemVerVersion(version)).toBe(true);